+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Unreleased
 - field.StreamField: Add exists flag
 - utils.find_iter: Fix searching with a single substring
 - lib.games.sierra: More TIM structures
+- lib.platforms.director: Add rip_dir_from_exe
 
 0.9.0 - 2021-01-14
 ==================